requirements.txt
streamlit==1.38.0
PyMuPDF==1.24.9         # for PDF text extraction
pdfminer.six==20231228   # fallback PDF extraction
python-docx==1.1.2       # DOCX text extraction
PyYAML==6.0.2            # read users.yaml
scikit-learn==1.5.2      # cosine similarity (optional nicety)
